fixing things, it works now no smoothing when creating route
